+++ lib/ocf_rpc.mli
@@ -1,0 +1,5 @@
+module V1 : sig
+  val halt : unit -> unit
+
+  val try_format : string -> string
+end